pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "life-insurance-assistant"
version = "1.0.0"
description = "AI-powered Life Insurance Support Assistant using LangGraph, Mem0, and configurable knowledge base"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Jaid"},
]

dependencies = [
    # Core AI/LLM
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-ollama>=0.2.0",
    "openai>=1.50.0",
    # Memory
    "mem0ai>=0.1.0",
    # Vector Store
    "qdrant-client>=1.9.0",
    "langchain-qdrant>=0.1.0",
    # API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    # CLI
    "rich>=13.0.0",
    "typer>=0.12.0",
    # Knowledge Base
    "pyyaml>=6.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
]

[project.scripts]
lia-cli = "src.cli.chat:app"
lia-server = "src.api.app:start_server"
lia-index = "src.knowledge.indexer:cli_index"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
